change busy-loop to double to allow for more granularity

+++ unit_tests/recovery/ut_redo_job_impl.hpp
@@ -48,7 +48,7 @@ class ut_redo_job_impl final : public cublog::redo_parallel::redo_job_base
 
   public:
     ut_redo_job_impl (ut_database &a_database_recovery, job_type a_job_type,
-		      INT64 a_id, VPID a_vpid, int a_millis);
+		      INT64 a_id, VPID a_vpid, double a_millis);
 
     ut_redo_job_impl (ut_redo_job_impl const &) = delete;
     ut_redo_job_impl (ut_redo_job_impl &&) = delete;
@@ -72,7 +72,7 @@ class ut_redo_job_impl final : public cublog::redo_parallel::redo_job_base
      */
     ux_ut_redo_job_impl clone ();
 
-    static void busy_loop (size_t a_millis);
+    static void busy_loop (double a_millis);
 
   private:
     ut_database &m_database_recovery;
@@ -80,7 +80,7 @@ class ut_redo_job_impl final : public cublog::redo_parallel::redo_job_base
     const job_type m_job_type;
 
     const INT64 m_id;
-    const int m_millis;
+    const double m_millis;
 };
 
 #endif // ! _UT_REDO_JOB_IMPL_H_
